RedisRepository classes are injecting as scoped
To resolve the multi-thread issue for the Redis repository class, the classes
inherited by RedisRepositoryBase are injecting as scoped to DI. Before that, they were injecting as a singleton and it was causing a mixing
threading issue for the UpdateBranches method. The developers want to
inject  Redis Repository classes as a singleton, must take care of the
threading issues. They can be easily solved by wrapping overriding
RedisRepositoryBase methods and wrapping UpdateBranches method
with the lock keyword.